Invoking Single Responsibility
- Add new file Persisitance.swift in Storage.
- Define a new struct called PersistanceController and declare a static constant of a shared instance.
- Initialize a persistance container which takes in a inMemory parameter for showing fake data in SwiftUI Previews.
- Fix typo on Persistance.swift to Persistence.swift.
- Add a static variable preview for a in-memory implementation of the persistence controller.
- Add static variable previewItem of type ExpenseModel.
- Remove static variable container, previewContainer and previewItem in AppMain and updates instances to use PersistenceController.
- Remove unused import CoreData.
