+++ ExpenseTracker/Storage/MonthlyReportsDataSource.swift
@@ -38,7 +38,7 @@ class MonthlyReportsDataSource: ObservableObject {
 
   @Published private(set) var currentEntries: [ExpenseModel] = []
 
-  init(viewContext: NSManagedObjectContext = AppMain.container.viewContext) {
+  init(viewContext: NSManagedObjectContext = PersistenceController.shared.container.viewContext) {
     self.viewContext = viewContext
     prepare()
   }
