+++ ExpenseTracker/Views/MonthlyExpensesView.swift
@@ -67,7 +67,7 @@ struct MonthlyExpensesView: View {
 
 struct MonthlyExpensesView_Previews: PreviewProvider {
   static var previews: some View {
-    let reportsDataSource = MonthlyReportsDataSource(viewContext: AppMain.previewContainer.viewContext)
+    let reportsDataSource = MonthlyReportsDataSource(viewContext: PersistenceController.preview.container.viewContext)
     MonthlyExpensesView(dataSource: reportsDataSource)
   }
 }
